source code/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.7.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cvlib==0.2.7
dm-tree==0.1.8
Flask==3.0.2
flatbuffers==24.3.7
gast==0.5.4
google-pasta==0.2.0
googletrans==4.0.0rc1
grpcio==1.62.1
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.34.0
importlib-metadata==7.0.1
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.3
keras==3.0.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
progressbar==2.5
protobuf==4.25.3
Pygments==2.17.2
pytesseract==0.3.10
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
six==1.16.0
sniffio==1.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
